fix: refactor docker_manager
